Global context: exposed config (ServerOpts)
